move validateDTO to the Package domain class
